feat(plugin-git): improve performance and tweaks options
