+++ plugins/development/plugin-git/src/node/index.ts
@@ -1,3 +1,6 @@
 export * from './gitPlugin.js'
-export type * from './types.js'
+export * from './resolveChangelog.js'
+export * from './resolveContributors.js'
 export * from './utils/index.js'
+export type * from './options.js'
+export type * from './typings.js'
