+++ plugins/development/plugin-git/src/node/utils/digestSHA256.ts
@@ -1,0 +1,9 @@
+import { createHash } from 'node:crypto'
+
+const hash = createHash('sha256')
+
+export const digestSHA256 = (message: string): string => {
+  hash.update(message)
+
+  return hash.digest('hex')
+}
