+++ plugins/development/plugin-git/src/node/utils/getUserNameWithNoreplyEmail.ts
@@ -1,0 +1,4 @@
+export const getUserNameWithNoreplyEmail = (email: string): string | null =>
+  email.endsWith('@users.noreply.github.com')
+    ? email.replace('@users.noreply.github.com', '').split('+')[1]
+    : null
