+++ plugins/development/plugin-git/src/node/utils/index.ts
@@ -1,4 +1,6 @@
 export * from './checkGitRepo.js'
+export * from './digestSHA256.js'
 export * from './getCommits.js'
-export * from './resolveContributors.js'
-export * from './resolveChangelog.js'
+export * from './getContributorInfo.js'
+export * from './getUserNameWithNoreplyEmail.js'
+export * from './inferGitProvider.js'
